LPS-82444 Improve logic

+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/internal/portlet/JournalPortlet.java
@@ -72,6 +72,7 @@ import com.liferay.journal.web.configuration.JournalWebConfiguration;
 import com.liferay.journal.web.internal.portlet.action.ActionUtil;
 import com.liferay.journal.web.util.JournalPortletUtil;
 import com.liferay.journal.web.util.JournalUtil;
+import com.liferay.layout.page.template.service.LayoutPageTemplateEntryLocalService;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.diff.CompareVersionsException;
@@ -1488,14 +1489,22 @@ public class JournalPortlet extends MVCPortlet {
 			_assetDisplayPageEntryLocalService.fetchAssetDisplayPageEntry(
 				groupId, classNameId, classPK);
 
-		if (assetDisplayPageEntry != null) {
-			_assetDisplayPageEntryLocalService.deleteAssetDisplayPageEntry(
-				groupId, classNameId, classPK);
+		if (displayPageType == AssetDisplayPageConstants.TYPE_NONE) {
+			if (assetDisplayPageEntry != null) {
+				_assetDisplayPageEntryLocalService.deleteAssetDisplayPageEntry(
+					groupId, classNameId, classPK);
+			}
 		}
-
-		_assetDisplayPageEntryLocalService.addAssetDisplayPageEntry(
-			userId, groupId, classNameId, classPK, assetDisplayPageId,
-			displayPageType, serviceContext);
+		else if (assetDisplayPageEntry == null) {
+			_assetDisplayPageEntryLocalService.addAssetDisplayPageEntry(
+				userId, groupId, classNameId, classPK, assetDisplayPageId,
+				displayPageType, serviceContext);
+		}
+		else {
+			_assetDisplayPageEntryLocalService.updateAssetDisplayPageEntry(
+				assetDisplayPageEntry.getAssetDisplayPageEntryId(),
+				assetDisplayPageId, displayPageType);
+		}
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(JournalPortlet.class);
@@ -1541,6 +1550,10 @@ public class JournalPortlet extends MVCPortlet {
 
 	@Reference
 	private LayoutLocalService _layoutLocalService;
+
+	@Reference
+	private LayoutPageTemplateEntryLocalService
+		_layoutPageTemplateEntryLocalService;
 
 	@Reference
 	private Portal _portal;
